+++ lib/src/widgets/help_button_widget.dart
@@ -6,17 +6,25 @@ class MPHelpButtonWidget extends StatelessWidget {
   final BuildContext context;
   final String helpPage;
   final String title;
+  final IconData? iconData;
+  final String? tooltip;
 
-  MPHelpButtonWidget(this.context, this.helpPage, this.title);
+  MPHelpButtonWidget(
+    this.context,
+    this.helpPage,
+    this.title, {
+    this.iconData,
+    this.tooltip,
+  });
 
   @override
   Widget build(BuildContext context) {
     return IconButton(
       key: ValueKey('$helpPage HelpPageButton'),
-      icon: Icon(Icons.help_outline),
+      icon: Icon(iconData ?? Icons.help_outline),
       color: Theme.of(context).colorScheme.onSecondaryContainer,
       onPressed: () => MPDialogAux.showHelpDialog(context, helpPage, title),
-      tooltip: AppLocalizations.of(context).helpDialogTooltip,
+      tooltip: tooltip ?? AppLocalizations.of(context).helpDialogTooltip,
     );
   }
 }